Renamed 'Beacon' -> 'AnalyticsBase'
